+++ src/writer.rs
@@ -1,10 +1,11 @@
 use std::fs::File;
 use std::io::{BufWriter, Result, Write};
 use std::path::Path;
+use std::vec;
 
 use super::Object::*;
 use super::{Dictionary, Document, Object, Stream, StringFormat};
-use crate::xref::*;
+use crate::{xref::*, IncrementalDocument};
 
 impl Document {
     /// Save PDF document to specified file path.
@@ -27,7 +28,7 @@ impl Document {
             bytes_written: 0,
         };
 
-        let mut xref = Xref::new(self.max_id + 1);
+        let mut xref = Xref::new(self.max_id + 1, self.reference_table.cross_reference_type);
         writeln!(target, "%PDF-{}", self.version)?;
 
         for (&(id, generation), object) in &self.objects {
@@ -42,9 +43,67 @@ impl Document {
         }
 
         let xref_start = target.bytes_written;
-        Writer::write_xref(&mut target, &xref)?;
-        self.write_trailer(&mut target)?;
+
+        // Pick right cross reference stream.
+        match xref.cross_reference_type {
+            XrefType::CrossReferenceTable => {
+                Writer::write_xref(&mut target, &xref)?;
+                self.write_trailer(&mut target)?;
+            }
+            XrefType::CrossReferenceStream => {
+                // Cross Reference Stream instead of XRef and Trailer
+                self.write_cross_reference_stream(&mut target, &mut xref, xref_start as u32)?;
+            }
+        }
+        // Write `startxref` part of trailer
         write!(target, "\nstartxref\n{}\n%%EOF", xref_start)?;
+
+        Ok(())
+    }
+
+    /// Write the Cross Reference Stream.
+    ///
+    /// Insert an `Object` to the end of the PDF (not visible when inspecting `Document`).
+    /// Note: This is different from the "Cross Reference Table".
+    fn write_cross_reference_stream<W: Write>(
+        &mut self, file: &mut CountingWrite<&mut W>, xref: &mut Xref, xref_start: u32,
+    ) -> Result<()> {
+        // Increment max_id to account for CRS.
+        self.max_id += 1;
+        let new_obj_id_for_crs = self.max_id;
+        xref.insert(
+            new_obj_id_for_crs,
+            XrefEntry::Normal {
+                offset: xref_start,
+                generation: 0,
+            },
+        );
+        // Update `max_id` in trailer
+        self.trailer.set("Size", i64::from(self.max_id + 1));
+        // Set the size of each entry in bytes (default for PDFs is `[1 2 1]`)
+        // In our case we use `[u8, u32, u16]` for each entry
+        // to keep things simple and working at all times.
+        self.trailer.set("W", Array(vec![Integer(1), Integer(4), Integer(2)]));
+        // Note that `ASCIIHexDecode` does not work correctly,
+        // but is still useful for debugging sometimes.
+        let filter = XRefStreamFilter::None;
+        let (stream, stream_length, indexes) = Writer::create_xref_steam(xref, filter)?;
+        self.trailer.set("Index", indexes);
+        if filter == XRefStreamFilter::ASCIIHexDecode {
+            self.trailer.set("Filter", Name(b"ASCIIHexDecode".to_vec()));
+        }
+        self.trailer.set("Length", stream_length as i64);
+
+        let trailer = &self.trailer;
+        let cross_reference_stream = Stream(Stream {
+            dict: trailer.clone(),
+            allows_compression: true,
+            content: stream,
+            start_position: None,
+        });
+        // Insert Cross Reference Stream as an `Object` to the end of the PDF.
+        // The `Object` is not added to `Document` because it is generated every time you save.
+        Writer::write_indirect_object(file, new_obj_id_for_crs, 0, &cross_reference_stream, xref)?;
 
         Ok(())
     }
@@ -57,7 +116,85 @@ impl Document {
     }
 }
 
+impl IncrementalDocument {
+    /// Save PDF document to specified file path.
+    #[inline]
+    pub fn save<P: AsRef<Path>>(&mut self, path: P) -> Result<File> {
+        let mut file = BufWriter::new(File::create(path)?);
+        self.save_internal(&mut file)?;
+        Ok(file.into_inner()?)
+    }
+
+    /// Save PDF to arbitrary target
+    #[inline]
+    pub fn save_to<W: Write>(&mut self, target: &mut W) -> Result<()> {
+        self.save_internal(target)
+    }
+
+    fn save_internal<W: Write>(&mut self, target: &mut W) -> Result<()> {
+        let mut target = CountingWrite {
+            inner: target,
+            bytes_written: 0,
+        };
+
+        // Write previous document versions.
+        let prev_document_bytes = self.get_prev_documents_bytes();
+        target.inner.write_all(prev_document_bytes)?;
+        target.bytes_written += prev_document_bytes.len();
+
+        // Write/Append new document version.
+        let mut xref = Xref::new(
+            self.new_document.max_id + 1,
+            self.get_prev_documents().reference_table.cross_reference_type,
+        );
+
+        if let Some(last_byte) = prev_document_bytes.last() {
+            if *last_byte != b'\n' {
+                // Add a newline if it was not already present
+                writeln!(target)?;
+            }
+        }
+        writeln!(target, "%PDF-{}", self.new_document.version)?;
+
+        for (&(id, generation), object) in &self.new_document.objects {
+            if object
+                .type_name()
+                .map(|name| ["ObjStm", "XRef", "Linearized"].contains(&name))
+                .ok()
+                != Some(true)
+            {
+                Writer::write_indirect_object(&mut target, id, generation, object, &mut xref)?;
+            }
+        }
+
+        let xref_start = target.bytes_written;
+
+        // Pick right cross reference stream.
+        match xref.cross_reference_type {
+            XrefType::CrossReferenceTable => {
+                Writer::write_xref(&mut target, &xref)?;
+                self.new_document.write_trailer(&mut target)?;
+            }
+            XrefType::CrossReferenceStream => {
+                // Cross Reference Stream instead of XRef and Trailer
+                self.new_document
+                    .write_cross_reference_stream(&mut target, &mut xref, xref_start as u32)?;
+            }
+        }
+        // Write `startxref` part of trailer
+        write!(target, "\nstartxref\n{}\n%%EOF", xref_start)?;
+
+        Ok(())
+    }
+}
+
 pub struct Writer;
+
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+pub enum XRefStreamFilter {
+    ASCIIHexDecode,
+    None,
+}
 
 impl Writer {
     fn need_separator(object: &Object) -> bool {
@@ -71,25 +208,128 @@ impl Writer {
         )
     }
 
+    /// Write Cross Reference Table.
+    ///
+    /// Note: This is different from a "Cross Reference Stream".
     fn write_xref(file: &mut dyn Write, xref: &Xref) -> Result<()> {
-        writeln!(file, "xref\n0 {}", xref.size)?;
+        writeln!(file, "xref")?;
 
-        let mut write_xref_entry =
-            |offset: u32, generation: u16, kind: char| writeln!(file, "{:>010} {:>05} {} ", offset, generation, kind);
-        write_xref_entry(0, 65535, 'f')?;
+        let mut xref_section = XrefSection::new(0);
+        // Add first (0) entry
+        xref_section.add_unusable_free_entry();
 
-        let mut obj_id = 1;
-        while obj_id < xref.size {
-            if let Some(entry) = xref.get(obj_id) {
-                if let XrefEntry::Normal { offset, generation } = *entry {
-                    write_xref_entry(offset, generation, 'n')?;
-                };
-            } else {
-                write_xref_entry(0, 65535, 'f')?;
+        for obj_id in 1..xref.size {
+            // If section is empty change number of starting id.
+            if xref_section.is_empty() {
+                xref_section = XrefSection::new(obj_id);
             }
-            obj_id += 1;
+            if let Some(entry) = xref.get(obj_id) {
+                match *entry {
+                    XrefEntry::Normal { offset, generation } => {
+                        // Add entry
+                        xref_section.add_entry(XrefEntry::Normal { offset, generation });
+                    }
+                    XrefEntry::Compressed { container: _, index: _ } => {
+                        xref_section.add_unusable_free_entry();
+                    }
+                    XrefEntry::Free => {
+                        xref_section.add_entry(XrefEntry::Free);
+                    }
+                    XrefEntry::UnusableFree => {
+                        xref_section.add_unusable_free_entry();
+                    }
+                }
+            } else {
+                // Skip over `obj_id`, but finish section if not empty.
+                if !xref_section.is_empty() {
+                    xref_section.write_xref_section(file)?;
+                    xref_section = XrefSection::new(obj_id);
+                }
+            }
+        }
+        // Print last section
+        if !xref_section.is_empty() {
+            xref_section.write_xref_section(file)?;
         }
         Ok(())
+    }
+
+    /// Create stream for Cross reference stream.
+    fn create_xref_steam(xref: &Xref, filter: XRefStreamFilter) -> Result<(Vec<u8>, usize, Object)> {
+        let mut xref_sections = Vec::new();
+        let mut xref_section = XrefSection::new(0);
+
+        for obj_id in 1..xref.size + 1 {
+            // If section is empty change number of starting id.
+            if xref_section.is_empty() {
+                xref_section = XrefSection::new(obj_id);
+            }
+            if let Some(entry) = xref.get(obj_id) {
+                xref_section.add_entry(entry.clone());
+            } else {
+                // Skip over but finish section if not empty
+                if !xref_section.is_empty() {
+                    xref_sections.push(xref_section);
+                    xref_section = XrefSection::new(obj_id);
+                }
+            }
+        }
+        // Print last section
+        if !xref_section.is_empty() {
+            xref_sections.push(xref_section);
+        }
+
+        let mut xref_stream = Vec::new();
+        let mut xref_index = Vec::new();
+
+        for section in xref_sections {
+            // Add indexes to list
+            xref_index.push(Integer(section.starting_id as i64));
+            xref_index.push(Integer(section.entries.len() as i64));
+            // Add entries to stream
+            let mut obj_id = section.starting_id;
+            for entry in section.entries {
+                match entry {
+                    XrefEntry::Free => {
+                        // Type 0
+                        xref_stream.push(0); // Type 0
+                        xref_stream.extend(obj_id.to_be_bytes());
+                        xref_stream.extend(vec![0, 0]); // TODO add generation number
+                    }
+                    XrefEntry::UnusableFree => {
+                        // Type 0
+                        xref_stream.push(0); // Type 0
+                        xref_stream.extend(obj_id.to_be_bytes());
+                        xref_stream.extend(65535_u16.to_be_bytes());
+                    }
+                    XrefEntry::Normal { offset, generation } => {
+                        // Type 1
+                        xref_stream.push(1); // Type 1
+                        xref_stream.extend(offset.to_be_bytes());
+                        xref_stream.extend(generation.to_be_bytes());
+                    }
+                    XrefEntry::Compressed { container, index } => {
+                        // Type 2
+                        xref_stream.push(2); // Type 2
+                        xref_stream.extend(container.to_be_bytes());
+                        xref_stream.extend(index.to_be_bytes());
+                    }
+                }
+                obj_id += 1;
+            }
+        }
+
+        // The end of line character should not be counted, added later.
+        let stream_length = xref_stream.len();
+
+        if filter == XRefStreamFilter::ASCIIHexDecode {
+            xref_stream = xref_stream
+                .iter()
+                .flat_map(|c| format!("{:02X}", c).as_bytes().to_vec())
+                .collect::<Vec<u8>>();
+        }
+
+        Ok((xref_stream, stream_length, Array(xref_index)))
     }
 
     fn write_indirect_object<W: Write>(
@@ -99,7 +339,7 @@ impl Writer {
         xref.insert(id, XrefEntry::Normal { offset, generation });
         write!(
             file,
-            "{} {} obj{}",
+            "{} {} obj\n{}",
             id,
             generation,
             if Writer::need_separator(object) { " " } else { "" }
@@ -107,7 +347,7 @@ impl Writer {
         Writer::write_object(file, object)?;
         writeln!(
             file,
-            "{}endobj",
+            "{}\nendobj",
             if Writer::need_end_separator(object) { " " } else { "" }
         )?;
         Ok(())
@@ -235,7 +475,7 @@ impl Writer {
         Writer::write_dictionary(file, &stream.dict)?;
         file.write_all(b"stream\n")?;
         file.write_all(&stream.content)?;
-        file.write_all(b"endstream")?;
+        file.write_all(b"\nendstream")?;
         Ok(())
     }
 }